YES I FIXED IT
stayed up till 1130 to find the error. The error was:

chunk checking was in a right to left motion, so the balls would get shifted to the left. This was why there was a natural bias towards the balls getting put into the wall on the left.

To fix this. I hold a temporary variable that holds the next pos and vec of the ball without directly changing it. Then after all the collisions, I change the actual pos and vec.